ConfigurationProcessor/tools/regress/requirements.txt
json-delta==1.1.3
pyyaml
sh
simplejson
six
termcolor
